Update Introductory Problems_Weird Algorithm.c

+++ Introductory Problems_Weird Algorithm.c
@@ -2,15 +2,15 @@
 #include <stdio.h>
 
 int main() {
-    int N;
-    scanf("%d", &N);
-    printf("%d ", N);
+    long long N;
+    scanf("%lld", &N);
+    printf("%lld ", N);
     while(N!=1){
         if(N%2 == 0){
             N = N/2;
         }
         else
             N = N*3 +1;
-        printf("%d ", N);
+        printf("%lld ", N);
     }
 }
